Added factorial function. Refactored unit tests to use counter instead of bool.
When running all the unit test functions with boolean values the unit tests stop running after the first function with a failed test. Changing to adding the number of failed tests fixes this.

As an aside... an int32_t will only hold up to factorial(12). A uint64_t will only hold up to factorial(20). Hardly worth the effort of calculating! Could have saved myself some trouble with a lookup table.